update blueprint ergonomics
