+++ blueprint/billing-base/api/controllers/checkoutSession.controller.ts
@@ -1,7 +1,7 @@
 import {
   handlers,
   IdSchema,
-  SchemaValidator,
+  schemaValidator,
   string
 } from '@forklaunch/blueprint-core';
 import { Metrics } from '@forklaunch/blueprint-monitoring';
@@ -23,7 +23,7 @@ export const CheckoutSessionController = (
 ) =>
   ({
     createCheckoutSession: handlers.post(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'createCheckoutSession',
@@ -42,7 +42,7 @@ export const CheckoutSessionController = (
     ),
 
     getCheckoutSession: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'getCheckoutSession',
@@ -61,7 +61,7 @@ export const CheckoutSessionController = (
     ),
 
     expireCheckoutSession: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id/expire',
       {
         name: 'expireCheckoutSession',
@@ -79,7 +79,7 @@ export const CheckoutSessionController = (
     ),
 
     handleCheckoutSuccess: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id/success',
       {
         name: 'handleCheckoutSuccess',
@@ -99,7 +99,7 @@ export const CheckoutSessionController = (
     ),
 
     handleCheckoutFailure: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id/failure',
       {
         name: 'handleCheckoutFailure',
