+++ blueprint/billing-base/api/controllers/paymentLink.controller.ts
@@ -3,7 +3,7 @@ import {
   handlers,
   IdSchema,
   IdsSchema,
-  SchemaValidator,
+  schemaValidator,
   string
 } from '@forklaunch/blueprint-core';
 import { Metrics } from '@forklaunch/blueprint-monitoring';
@@ -26,7 +26,7 @@ export const PaymentLinkController = (
 ) =>
   ({
     createPaymentLink: handlers.post(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'createPaymentLink',
@@ -45,7 +45,7 @@ export const PaymentLinkController = (
     ),
 
     getPaymentLink: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'getPaymentLink',
@@ -62,7 +62,7 @@ export const PaymentLinkController = (
     ),
 
     updatePaymentLink: handlers.put(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'updatePaymentLink',
@@ -82,7 +82,7 @@ export const PaymentLinkController = (
     ),
 
     expirePaymentLink: handlers.delete(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'expirePaymentLink',
@@ -100,7 +100,7 @@ export const PaymentLinkController = (
     ),
 
     handlePaymentSuccess: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id/success',
       {
         name: 'handlePaymentSuccess',
@@ -121,7 +121,7 @@ export const PaymentLinkController = (
     ),
 
     handlePaymentFailure: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id/failure',
       {
         name: 'handlePaymentFailure',
@@ -142,7 +142,7 @@ export const PaymentLinkController = (
     ),
 
     listPaymentLinks: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'listPaymentLinks',
