+++ blueprint/billing-base/api/controllers/plan.controller.ts
@@ -3,7 +3,7 @@ import {
   handlers,
   IdSchema,
   IdsSchema,
-  SchemaValidator,
+  schemaValidator,
   string
 } from '@forklaunch/blueprint-core';
 import { Metrics } from '@forklaunch/blueprint-monitoring';
@@ -20,14 +20,13 @@ import {
 } from '../../domain/mappers/plan.mappers';
 import { PlanSchemas } from '../../registrations';
 import { PlanServiceFactory } from '../routes/plan.routes';
-
 export const PlanController = (
   serviceFactory: PlanServiceFactory,
   openTelemetryCollector: OpenTelemetryCollector<Metrics>
 ) =>
   ({
     createPlan: handlers.post(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'createPlan',
@@ -44,7 +43,7 @@ export const PlanController = (
     ),
 
     getPlan: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'getPlan',
@@ -65,7 +64,7 @@ export const PlanController = (
     ),
 
     updatePlan: handlers.put(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'updatePlan',
@@ -82,7 +81,7 @@ export const PlanController = (
     ),
 
     deletePlan: handlers.delete(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'deletePlan',
@@ -100,7 +99,7 @@ export const PlanController = (
     ),
 
     listPlans: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'listPlans',
