+++ blueprint/billing-base/api/routes/subscription.routes.ts
@@ -1,9 +1,8 @@
-import { forklaunchRouter, SchemaValidator } from '@forklaunch/blueprint-core';
+import { forklaunchRouter, schemaValidator } from '@forklaunch/blueprint-core';
 import { sdkRouter } from '@forklaunch/core/http';
 import { ci, tokens } from '../../server';
 import { SubscriptionController } from '../controllers/subscription.controller';
 
-const schemaValidator = SchemaValidator();
 const openTelemetryCollector = ci.resolve(tokens.OpenTelemetryCollector);
 const subscriptionServiceFactory = ci.scopedResolver(
   tokens.SubscriptionService
