--- blueprint/billing-base/sdk.types.ts
@@ -1,3 +1,0 @@
-import { billingSdk } from './server';
-
-export type BillingSdkClient = typeof billingSdk;
