+++ blueprint/billing-stripe/api/controllers/plan.controller.ts
@@ -3,7 +3,7 @@ import {
   handlers,
   IdSchema,
   IdsSchema,
-  SchemaValidator,
+  schemaValidator,
   string
 } from '@forklaunch/blueprint-core';
 import { Metrics } from '@forklaunch/blueprint-monitoring';
@@ -29,7 +29,7 @@ export const PlanController = (
 ) =>
   ({
     createPlan: handlers.post(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'createPlan',
@@ -46,7 +46,7 @@ export const PlanController = (
     ),
 
     getPlan: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'getPlan',
@@ -63,7 +63,7 @@ export const PlanController = (
     ),
 
     updatePlan: handlers.put(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'updatePlan',
@@ -80,7 +80,7 @@ export const PlanController = (
     ),
 
     deletePlan: handlers.delete(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'deletePlan',
@@ -98,7 +98,7 @@ export const PlanController = (
     ),
 
     listPlans: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'listPlans',
