+++ blueprint/billing-stripe/api/controllers/subscription.controller.ts
@@ -3,7 +3,7 @@ import {
   handlers,
   IdSchema,
   IdsSchema,
-  SchemaValidator,
+  schemaValidator,
   string
 } from '@forklaunch/blueprint-core';
 import { Metrics } from '@forklaunch/blueprint-monitoring';
@@ -25,7 +25,7 @@ export const SubscriptionController = (
 ) =>
   ({
     createSubscription: handlers.post(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'createSubscription',
@@ -44,7 +44,7 @@ export const SubscriptionController = (
     ),
 
     getSubscription: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'getSubscription',
@@ -63,7 +63,7 @@ export const SubscriptionController = (
     ),
 
     getUserSubscription: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/user/:id',
       {
         name: 'getUserSubscription',
@@ -85,7 +85,7 @@ export const SubscriptionController = (
     ),
 
     getOrganizationSubscription: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/organization/:id',
       {
         name: 'getOrganizationSubscription',
@@ -107,7 +107,7 @@ export const SubscriptionController = (
     ),
 
     updateSubscription: handlers.put(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'updateSubscription',
@@ -127,7 +127,7 @@ export const SubscriptionController = (
     ),
 
     deleteSubscription: handlers.delete(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'deleteSubscription',
@@ -145,7 +145,7 @@ export const SubscriptionController = (
     ),
 
     listSubscriptions: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'listSubscriptions',
@@ -164,7 +164,7 @@ export const SubscriptionController = (
     ),
 
     cancelSubscription: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id/cancel',
       {
         name: 'cancelSubscription',
@@ -182,7 +182,7 @@ export const SubscriptionController = (
     ),
 
     resumeSubscription: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id/resume',
       {
         name: 'resumeSubscription',
