+++ blueprint/billing-stripe/registrations.ts
@@ -1,6 +1,7 @@
 import {
   number,
   optional,
+  schemaValidator,
   SchemaValidator,
   string
 } from '@forklaunch/blueprint-core';
@@ -67,23 +68,23 @@ import mikroOrmOptionsConfig from './mikro-orm.config';
 
 //! defines the schemas for the billing portal service
 export const BillingPortalSchemas = StripeBillingPortalServiceSchemas({
-  validator: SchemaValidator()
+  validator: schemaValidator
 });
 export const CheckoutSessionSchemas = StripeCheckoutSessionServiceSchemas({
-  validator: SchemaValidator()
+  validator: schemaValidator
 });
 export const PaymentLinkSchemas = StripePaymentLinkServiceSchemas({
-  validator: SchemaValidator()
+  validator: schemaValidator
 });
 export const PlanSchemas = StripePlanServiceSchemas({
-  validator: SchemaValidator()
+  validator: schemaValidator
 });
 export const SubscriptionSchemas = StripeSubscriptionServiceSchemas({
-  validator: SchemaValidator()
+  validator: schemaValidator
 });
 
 //! defines the configuration schema for the application
-const configInjector = createConfigInjector(SchemaValidator(), {
+const configInjector = createConfigInjector(schemaValidator, {
   SERVICE_METADATA: {
     lifetime: Lifetime.Singleton,
     type: {
@@ -209,7 +210,7 @@ const serviceDependencies = runtimeDependencies.chain({
           : EntityManager,
         TtlCache,
         OpenTelemetryCollector,
-        SchemaValidator(),
+        schemaValidator,
         {
           BillingPortalMapper,
           CreateBillingPortalMapper,
@@ -236,7 +237,7 @@ const serviceDependencies = runtimeDependencies.chain({
           : EntityManager,
         TtlCache,
         OpenTelemetryCollector,
-        SchemaValidator(),
+        schemaValidator,
         {
           CheckoutSessionMapper,
           CreateCheckoutSessionMapper,
@@ -263,7 +264,7 @@ const serviceDependencies = runtimeDependencies.chain({
           : EntityManager,
         TtlCache,
         OpenTelemetryCollector,
-        SchemaValidator(),
+        schemaValidator,
         {
           PaymentLinkMapper,
           CreatePaymentLinkMapper,
@@ -285,7 +286,7 @@ const serviceDependencies = runtimeDependencies.chain({
           ? resolve('EntityManager', context)
           : EntityManager,
         OpenTelemetryCollector,
-        SchemaValidator(),
+        schemaValidator,
         {
           PlanMapper,
           CreatePlanMapper,
@@ -311,7 +312,7 @@ const serviceDependencies = runtimeDependencies.chain({
           ? resolve('EntityManager', context)
           : EntityManager,
         OpenTelemetryCollector,
-        SchemaValidator(),
+        schemaValidator,
         {
           SubscriptionMapper,
           CreateSubscriptionMapper,
@@ -350,7 +351,7 @@ const serviceDependencies = runtimeDependencies.chain({
         context.entityManagerOptions
           ? resolve('EntityManager', context)
           : EntityManager,
-        SchemaValidator(),
+        schemaValidator,
         OpenTelemetryCollector,
         BillingPortalService,
         CheckoutSessionService,
@@ -361,7 +362,8 @@ const serviceDependencies = runtimeDependencies.chain({
   }
 });
 
-export const createDependencies = (envFilePath: string) => ({
+//! validates the configuration and returns the dependencies for the application
+export const createDependencyContainer = (envFilePath: string) => ({
   ci: serviceDependencies.validateConfigSingletons(envFilePath),
   tokens: serviceDependencies.tokens()
 });
