+++ blueprint/iam-base/api/controllers/user.controller.ts
@@ -3,7 +3,7 @@ import {
   handlers,
   IdSchema,
   IdsSchema,
-  SchemaValidator,
+  schemaValidator,
   string
 } from '@forklaunch/blueprint-core';
 import { Metrics } from '@forklaunch/blueprint-monitoring';
@@ -23,7 +23,7 @@ export const UserController = (
 ) =>
   ({
     createUser: handlers.post(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'Create User',
@@ -43,7 +43,7 @@ export const UserController = (
     ),
 
     createBatchUsers: handlers.post(
-      SchemaValidator(),
+      schemaValidator,
       '/batch',
       {
         name: 'Create Batch Users',
@@ -62,7 +62,7 @@ export const UserController = (
     ),
 
     getUser: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'Get User',
@@ -80,7 +80,7 @@ export const UserController = (
     ),
 
     getBatchUsers: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/batch',
       {
         name: 'Get Batch Users',
@@ -98,7 +98,7 @@ export const UserController = (
     ),
 
     updateUser: handlers.put(
-      SchemaValidator(),
+      schemaValidator,
       '/',
       {
         name: 'Update User',
@@ -117,7 +117,7 @@ export const UserController = (
     ),
 
     updateBatchUsers: handlers.put(
-      SchemaValidator(),
+      schemaValidator,
       '/batch',
       {
         name: 'Update Batch Users',
@@ -136,7 +136,7 @@ export const UserController = (
     ),
 
     deleteUser: handlers.delete(
-      SchemaValidator(),
+      schemaValidator,
       '/:id',
       {
         name: 'Delete User',
@@ -155,7 +155,7 @@ export const UserController = (
     ),
 
     deleteBatchUsers: handlers.delete(
-      SchemaValidator(),
+      schemaValidator,
       '/batch',
       {
         name: 'Delete Batch Users',
@@ -174,7 +174,7 @@ export const UserController = (
     ),
 
     verifyHasRole: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id/verify-role/:roleId',
       {
         name: 'Verify User Role',
@@ -197,7 +197,7 @@ export const UserController = (
     ),
 
     verifyHasPermission: handlers.get(
-      SchemaValidator(),
+      schemaValidator,
       '/:id/verify-permission/:permissionId',
       {
         name: 'Verify User Permission',
