+++ blueprint/iam-base/api/routes/permission.routes.ts
@@ -1,9 +1,8 @@
-import { forklaunchRouter, SchemaValidator } from '@forklaunch/blueprint-core';
+import { forklaunchRouter, schemaValidator } from '@forklaunch/blueprint-core';
 import { sdkRouter } from '@forklaunch/core/http';
 import { ci, tokens } from '../../server';
 import { PermissionController } from '../controllers/permission.controller';
 
-const schemaValidator = SchemaValidator();
 const openTelemetryCollector = ci.resolve(tokens.OpenTelemetryCollector);
 const permissionServiceFactory = ci.scopedResolver(tokens.PermissionService);
