+++ blueprint/iam-base/registrations.ts
@@ -1,6 +1,7 @@
 import {
   number,
   optional,
+  schemaValidator,
   SchemaValidator,
   string
 } from '@forklaunch/blueprint-core';
@@ -56,22 +57,22 @@ import mikroOrmOptionsConfig from './mikro-orm.config';
 //! defines the schemas for the organization service
 export const OrganizationSchemas = BaseOrganizationServiceSchemas({
   uuidId: true,
-  validator: SchemaValidator()
+  validator: schemaValidator
 });
 export const PermissionSchemas = BasePermissionServiceSchemas({
   uuidId: true,
-  validator: SchemaValidator()
+  validator: schemaValidator
 });
 export const RoleSchemas = BaseRoleServiceSchemas({
   uuidId: true,
-  validator: SchemaValidator()
+  validator: schemaValidator
 });
 export const UserSchemas = BaseUserServiceSchemas({
   uuidId: true,
-  validator: SchemaValidator()
+  validator: schemaValidator
 });
 //! defines the configuration schema for the application
-const configInjector = createConfigInjector(SchemaValidator(), {
+const configInjector = createConfigInjector(schemaValidator, {
   SERVICE_METADATA: {
     lifetime: Lifetime.Singleton,
     type: {
@@ -166,7 +167,7 @@ const serviceDependencies = runtimeDependencies.chain({
           ? resolve('EntityManager', context)
           : EntityManager,
         OpenTelemetryCollector,
-        SchemaValidator(),
+        schemaValidator,
         {
           OrganizationMapper,
           CreateOrganizationMapper,
@@ -184,7 +185,7 @@ const serviceDependencies = runtimeDependencies.chain({
           : EntityManager,
         () => resolve('RoleService', context),
         OpenTelemetryCollector,
-        SchemaValidator(),
+        schemaValidator,
         {
           PermissionMapper,
           CreatePermissionMapper,
@@ -202,7 +203,7 @@ const serviceDependencies = runtimeDependencies.chain({
           ? resolve('EntityManager', context)
           : EntityManager,
         OpenTelemetryCollector,
-        SchemaValidator(),
+        schemaValidator,
         {
           RoleMapper,
           CreateRoleMapper,
@@ -232,7 +233,7 @@ const serviceDependencies = runtimeDependencies.chain({
         () => resolve('RoleService', context),
         () => resolve('OrganizationService', context),
         OpenTelemetryCollector,
-        SchemaValidator(),
+        schemaValidator,
         {
           UserMapper,
           CreateUserMapper,
@@ -241,7 +242,9 @@ const serviceDependencies = runtimeDependencies.chain({
       )
   }
 });
-export const createDependencies = (envFilePath: string) => ({
+
+//! validates the configuration and returns the dependencies for the application
+export const createDependencyContainer = (envFilePath: string) => ({
   ci: serviceDependencies.validateConfigSingletons(envFilePath),
   tokens: serviceDependencies.tokens()
 });
