--- blueprint/iam-base/sdk.types.ts
@@ -1,3 +1,0 @@
-import { iamSdk } from './server';
-
-export type IamSdkClient = typeof iamSdk;
