--- blueprint/sample-worker/sdk.types.ts
@@ -1,3 +1,0 @@
-import { sampleWorkerSdk } from './server';
-
-export type SampleWorkerSdkClient = typeof sampleWorkerSdk;
